Work towards v1 of the parser
Adds potential specifications for the parsing language, a state machine to parse rules that will also be the core of the gen 2 parser, and a start to the manually written parser that will make up the first gen of the parsing language. The goal is to first parse the language manually, and then to see if we can use the parser's output to automatically convert a pglsf file to a parser. After that, I may add more features to the machine, but I am not sure yet. The goal of the parsing language is to be simple and readable.